update and test multipole_k_m_and_phi_m_from
