+++ test_autogalaxy/profiles/mass/total/test_power_law_multipole.py
@@ -1,5 +1,6 @@
-import autogalaxy as ag
 import pytest
+
+import autogalaxy as ag
 
 grid = ag.Grid2DIrregular([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
