+++ test_autogalaxy/test_convert.py
@@ -128,3 +128,25 @@ def test__shear_magnitude_and_angle_from():
 
     assert magnitude == pytest.approx(0.07071067811865, 1.0e-4)
     assert angle == pytest.approx(-22.5, 1.0e-4)
+
+def test__multipole_k_m_and_phi_m_from():
+
+    k_m, phi = ag.convert.multipole_k_m_and_phi_m_from(multipole_comps=(0.1, 0.0), m=1)
+
+    assert k_m == pytest.approx(0.1, 1e-3)
+    assert phi == pytest.approx(90.0, 1e-3)
+
+    k_m, phi = ag.convert.multipole_k_m_and_phi_m_from(multipole_comps=(0.0, 0.1), m=1)
+
+    assert k_m == pytest.approx(0.1, 1e-3)
+    assert phi == pytest.approx(0.0, 1e-3)
+
+    k_m, phi = ag.convert.multipole_k_m_and_phi_m_from(multipole_comps=(0.1, 0.0), m=2)
+
+    assert k_m == pytest.approx(0.1, 1e-3)
+    assert phi == pytest.approx(45.0, 1e-3)
+
+    k_m, phi = ag.convert.multipole_k_m_and_phi_m_from(multipole_comps=(-0.1, -0.1), m=2)
+
+    assert k_m == pytest.approx(0.1, 1e-3)
+    assert phi == pytest.approx(0.141421356, 1e-3)
